requirements.txt
Babel==2.9.1
blinker==1.4
click==8.0.4
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.3
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-WTF==1.0.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.0
MarkupSafe==2.1.1
passlib==1.7.4
picamera==1.13
pkg-resources==0.0.0
python-dotenv==0.20.0
pytz==2022.1
RPi.GPIO==0.7.1
speaklater==1.3
typing-extensions==4.1.1
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.7.0
